backend/requirements.txt
Django>=4.2
mysqlclient
PyJWT
djangorestframework
django-cors-headers
